Added Use-case document file
This document contains the use-cases for the project. As outlined in the Project Phases, the Protocol Design phase requires the collection and analysis of the use-cases. Please add use-cases as it becomes apparent and strictly follow the document rules outlined in said document. Rules suck, but these are here to help everyone develop together- no matter each of our programming  experience!