Initial user interface setup
The package and classes for the user interface has been setup. As outlined in the Requirements v1.0 (posted on the root directory README.md), OBSIDIAN_ROBOT will be a GUI based application and not console based like the Protocol test applications. Only the server *may* remain as console based, but that remains to be seen!