Major GUI implementation
The GUI has been implemented. GUI is not my strong point, but this is something I am happy with and can continue to evolve and be worked on. The logic that sits behind the GUI has been created in a seperate class and thus can be worked on independently of the GUI design. This makes the GUI a sort of skin to the application.